+++ Sources/LogConsole/SmartLogAggregator.cs
@@ -9,7 +9,8 @@ using System.Linq;
 namespace KSPDev.LogConsole {
 
 /// <summary>A log capturer that aggregates logs globally by the content.</summary>
-[PersistentFieldsFileAttribute("KSPDev/KSPDev.settings", "SmartLogAggregator")]
+[PersistentFieldsFileAttribute("KSPDev/LogConsole/Plugins/PluginData/settings.cfg",
+                               "SmartLogAggregator")]
 sealed class SmartLogAggregator : BaseLogAggregator {
   /// <summary>Log index used by smart logging.</summary>
   readonly Dictionary<int, LinkedListNode<LogRecord>> logRecordsIndex =