+++ Sources/LogConsole/CHANGELOG.md
@@ -1,2 +1,5 @@
+### KPSPDev_LogConsole v0.13.0 (January 6th, 2017):
+- [Change] Rename .settings files into .cfg, and move them into PluginData.
+
 ### KPSPDev_LogConsole v0.12.0 (January 3rd, 2017):
 - Initial CKAN version.